% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox-to-polygon.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{Convert a bounding box to a polygon}
\usage{
bbox_to_polygon(bbox)
}
\arguments{
\item{bbox}{A numeric matrix with two columns and two rows.
\code{bbox} represents the minimum and maximum coordinates.}
}
\value{
A rectangle polygon with the same size as \code{bbox}.
}
\description{
This function convert a bounding box to a polygon.
}
