% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-weights-epanechnikov.R
\name{calculate_weights_epanechnikov}
\alias{calculate_weights_epanechnikov}
\title{Epanechnikov kernel}
\usage{
calculate_weights_epanechnikov(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of weights of the kernel for each input points.
}
\description{
The kernel is one of the quadratic kernels used
in kernel density estimation.
}
\details{
Epanechnikov kernel is defined as follows:
\deqn{K(x) = \frac{3}{4}(1 - x^2) \quad \text{if} \quad |x| \leq 1\text{.}}
}
\examples{
x <- seq(-3, 3, 0.1)
y <- calculate_weights_epanechnikov(x)
plot(x, y, type = "l")
}
