% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-epanechnikov.R
\name{compute_epanechnikov}
\alias{compute_epanechnikov}
\title{Compute the Epanechnikov function}
\usage{
compute_epanechnikov(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of weights of the kernel for each input points.
}
\description{
The Epanechnikov function is one of the quadratic kernels used in
kernel density estimation. It is defined as follows:
\deqn{K(x) = \frac{3}{4}(1 - x^2) \quad \text{if} \quad |x| \leq 1\text{.}}
}
\examples{
x <- seq(-3, 3, 0.1)
y <- compute_epanechnikov(x)
plot(x, y, xlim = c(-3, 3), ylim = c(0, 1), type = "l")
}
