% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-gaussian.R
\name{compute_gaussian}
\alias{compute_gaussian}
\title{Compute the Gaussian function}
\usage{
compute_gaussian(x, sigma = 1)
}
\arguments{
\item{x}{A numric vector.}

\item{sigma}{The standard deviation of the Gaussian function.}
}
\value{
A numeric vector of weights of the kernel for each input points.
}
\description{
The Gaussian function is a bell-shaped function defined as follows:
\deqn{
  K(x) =
    \frac{1}{\sqrt{2\pi\sigma^2}}
    \exp\left( -\frac{x^2}{2\sigma^2} \right)\text{.}
}
}
\examples{
x <- seq(-3, 3, 0.1)
y <- compute_gaussian(x, sigma = 1)
plot(x, y, xlim = c(-3, 3), ylim = c(0, 1), type = "l")
}
