% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-bbox-to-coordinates.R
\name{convert_bbox}
\alias{convert_bbox}
\alias{convert_bbox_to_coordinates}
\alias{convert_bbox_to_polygon}
\title{Convert a bounding box to coordinates or a polygon}
\usage{
convert_bbox(bbox, output = c("coordinates", "polygon"), crs = NULL)

convert_bbox_to_coordinates(bbox)

convert_bbox_to_polygon(bbox, crs = NULL)
}
\arguments{
\item{bbox}{A numeric matrix with two columns and two rows.
\code{bbox} represents the minimum and maximum coordinates.
The bounding box can be created using the \code{create_bbox} function.}

\item{output}{A string specifying the output type. Either "coordinates" or
"polygon".}

\item{crs}{A string specifying the coordinate reference system. This is
only used when converting to a polygon.}
}
\value{
A \code{coordinates} object or a \code{sfc} polygon object.
}
\description{
This function converts a bounding box to either coordinates or a polygon.
}
\examples{
bbox <- create_bbox(center_lon = 136.9024,
                    center_lat =  35.1649,
                    width      = 0.10,
                    height     = 0.05)
convert_bbox_to_coordinates(bbox)
convert_bbox_to_polygon(bbox, crs = 4326)
}
