% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-coordinates.R
\name{create_coordinates}
\alias{create_coordinates}
\title{Create a collection of coordinates.}
\usage{
create_coordinates(...)
}
\arguments{
\item{...}{A sequence of x and y coordinates.}
}
\value{
A \code{coordinates} object which is a matrix with x and y columns.
}
\description{
This function creates a collection of coordinates from a sequence
x and y coordinates.
}
\examples{
# Create a `coordinates` object from a sequence of x and y coordinates
create_coordinates(1, 2, 3, 4)

# Create a `coordinates` object from an with a vector of x and y
# coordinates
create_coordinates(c(1, 2, 3, 4))
}
