% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-line-graph.R
\name{create_line_graph}
\alias{create_line_graph}
\title{Create a line graph from a network}
\usage{
create_line_graph(network)
}
\arguments{
\item{network}{A \code{road_network} or \code{segmented_network} object.}
}
\value{
A \code{igraph} object representing the midpoint graph.
}
\description{
This function creates a line graph where nodes represent the midpoints
of links in the input network.
}
\examples{
# Create a road network
network <- create_road_network(sample_roads)

# Plot the road network
plot(network$graph)

# Create a midpoint graph from a road network
graph <- create_line_graph(network)

# Plot the midpoint graph
plot(graph)
}
