% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-linestring.R
\name{create_linestring}
\alias{create_linestring}
\title{Create a linestring geometry}
\usage{
create_linestring(x = NULL, ...)
}
\arguments{
\item{x}{A numeric vector, or the first x coordinate.}

\item{...}{Additional numeric vectors.}
}
\value{
A \code{sf} linestring object.
}
\description{
This function takes a series of numeric arguments representing x and y
coordinates and creates a linestring geometry object.
}
\examples{
# Create a linestring from individual coordinates
create_linestring(0, 0, 3, 0, 4, 3)

# Create a linestring from a numeric vector
create_linestring(c(0, 0, 3, 0, 4, 3))
}
