% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-polygon.R
\name{create_polygon}
\alias{create_polygon}
\title{Create polygon from bounding box}
\usage{
create_polygon(bbox)
}
\arguments{
\item{bbox}{A numeric matrix with two columns and two rows.
\code{bbox} represents the minimum and maximum coordinates.}
}
\value{
A rectangle polygon with the same size as \code{bbox}.
}
\description{
This function creates a rectangle polygon from a bounding box.
}
