% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-road-network.R
\name{create_road_network}
\alias{create_road_network}
\title{Create a road network from roads}
\usage{
create_road_network(roads, directed = FALSE)
}
\arguments{
\item{roads}{A linestring object representing roads.}

\item{directed}{Logical indicating whether the road network is directed.}
}
\value{
A road network object.
}
\description{
This function constructs a road network from a set of roads. The road
network is represented nodes and links between nodes. Nodes are defined
as intersections between roads and endpoints of road segments.
}
\examples{
# Create the road network
road_network <- create_road_network(demo_roads)
road_network

# Plot the road network
plot(road_network)
}
