% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-segmented-network.R
\name{create_segmented_network}
\alias{create_segmented_network}
\title{Create a segmented road network}
\usage{
create_segmented_network(road_network, segment_length = 1)
}
\arguments{
\item{road_network}{A \code{road_network} object representing the input road
network.}

\item{segment_length}{A numeric value specifying the length of each segment.}
}
\value{
A \code{segmented_network} object.
}
\description{
This function creates a segmented road network by dividing each link of the
input road network into segments of a specified length.
}
\examples{
# Create a road network
road_network <- create_road_network(demo_roads)

# Create a segmented road network
segmented_network <- create_segmented_network(road_network)
segmented_network

# Plot the segmented road network
plot(segmented_network)
}
