% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo-roads.R
\docType{data}
\name{demo_roads}
\alias{demo_roads}
\title{Demo roads dataset}
\format{
A \code{sf} linestring object
}
\usage{
demo_roads
}
\description{
\code{demo_roads} a sample dataset representing simple roads.
It intended for demonstration and testing.
}
\examples{
demo_roads
}
\keyword{datasets}
