% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-roads.R
\name{download_roads}
\alias{download_roads}
\title{Download and process OpenStreetMap road data}
\usage{
download_roads(bbox, crop = TRUE)
}
\arguments{
\item{bbox}{A bounding box in the form of a matrix with two rows and two
columns.}

\item{crop}{A logical value indicating whether to crop the data to the
bounding box. Defaults to \code{TRUE}.}
}
\value{
A data frame containing OpenStreetMap road data.
}
\description{
This function downloads OpenStreetMap highway data within the specified
bounding box. It corrects the geometries of reversed one-way roads.
Optionally, it crops the data to the bounding box.
}
\examples{
\dontrun{
# Download road data
bbox <- create_bbox(
  north =  35.17377,
  south =  35.16377,
  east  = 136.91590,
  west  = 136.90090
)
roads <- download_roads(bbox)

# Plot the roads
plot(roads$geometry)
}
}
