% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-durations.R
\name{durations}
\alias{durations}
\alias{create_durations}
\title{Create durations}
\usage{
create_durations(duration_length, ...)
}
\arguments{
\item{duration_length}{The length of each duration.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A \code{durations} object.
}
\description{
This function creates a data frame of durations.
}
\examples{
# Create durations of 1 hour
create_durations("1 hour")

# Create durations of 30 minutes
create_durations("30 minutes")
}
