% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-road-endpoints.R
\name{extract_road_endpoints}
\alias{extract_road_endpoints}
\title{Extract road endpoints}
\usage{
extract_road_endpoints(roads)
}
\arguments{
\item{roads}{A linestring object representing roads. It should be have
a column named \code{road_id}.}
}
\value{
A points object representing road endpoints. Each endpoint
has the following columns:
\itemize{
\item origins: A list of road IDs that intersect at the endpoint.
\item n.overlaps: The number of roads that intersect at the endpoint.
}
}
\description{
This function extracts the endpoints of a set of roads. It identifies
and counts the roads that intersect at each endpoint.
}
