% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-road-intersections.R
\name{extract_road_intersections}
\alias{extract_road_intersections}
\title{Extract road intersections}
\usage{
extract_road_intersections(roads)
}
\arguments{
\item{roads}{A linestring object representing roads. It should be have
a column named \code{layer} and a column named \code{road_id}.}
}
\value{
A data frame with the following columns:
\itemize{
\item parent_road: A list of road IDs that intersect at the intersection.
\item num_overlaps: The number of roads that intersect at the
intersection.
}
}
\description{
This function identifies and extracts intersections between roads.
It considers only intersections where multiple roads on the same layer.
}
\examples{
# Extract road intersections
intersections <- extract_road_intersections(sample_roads)
intersections

# Plot the intersections
plot(sample_roads$geometry)
plot(intersections$geometry, pch = 16, col = "#E69F00", add = TRUE)
}
