% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-road-network-links.R
\name{extract_road_network_links}
\alias{extract_road_network_links}
\title{Extract road network links}
\usage{
extract_road_network_links(roads, nodes)
}
\arguments{
\item{roads}{A linestring object representing the roads.}

\item{nodes}{A point object representing the road network nodes.}
}
\value{
A linestring object representing the road network links.
}
\description{
This function extracts the road network links from a set of road geometries.
It splits the road geometries at the nodes and creates a new sf object
representing the road network links.
}
\seealso{
To create the nodes, use the \code{\link[=extract_road_network_nodes]{extract_road_network_nodes()}}
}
