% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-road-network-nodes.R
\name{extract_road_network_nodes}
\alias{extract_road_network_nodes}
\title{Extract road network nodes}
\usage{
extract_road_network_nodes(roads)
}
\arguments{
\item{roads}{A linestring object representing roads.}
}
\value{
A points object representing road network nodes.
}
\description{
This function extracts nodes from roads represented as linestring objects
downloaded from OpenStreetMap. Nodes are defined as intersections between
roads and endpoints of road segments.
}
