% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-segment-nodes.R
\name{extract_segment_nodes}
\alias{extract_segment_nodes}
\title{Extract segment nodes from network links}
\usage{
extract_segment_nodes(road_network, segment_length)
}
\arguments{
\item{road_network}{A \code{road_network} object created with \code{create_network()}.}

\item{segment_length}{The length of each segment to sample along the links.}
}
\value{
An \code{sf} object with the sampled points.
}
\description{
This function extracts nodes at regular intervals along each link in
a road network
}
