% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-segmented-network-links.R
\name{extract_segmented_network_links}
\alias{extract_segmented_network_links}
\title{Extract segmented network links from a road network}
\usage{
extract_segmented_network_links(
  road_network,
  segmented_network_nodes,
  tolerance = 1e-05
)
}
\arguments{
\item{road_network}{A \code{road_network} object created with
\code{create_road_network()}}

\item{segmented_network_nodes}{A data frame representing the segmented
network nodes.}

\item{tolerance}{A numeric value representing the maximum distance allowed
between a linestring of a road link and a point of a segmented network
node. If the distance exceeds this value, the point will not be used to
split the linestring.}
}
\value{
A data frame representing the segmented network links.
}
\description{
This function extracts the segmented network links from a road network
based on a specified set of segmented network nodes. It splits the
linestrings of the road network links at the points of the segmented
network nodes to create the segmented network links.
}
\examples{
# Create a road network
road_network <- create_road_network(sample_roads)

# Extract segmented network nodes by length of 1
segmented_network_nodes <- extract_segmented_network_nodes(road_network, 1)

# Extract segmented network links
segmented_network_links <- extract_segmented_network_links(
  road_network,
  segmented_network_nodes
)
segmented_network_links

# Plot the segmented network nodes and links
plot(segmented_network_links$geometry, col = "#E69F00")
plot(segmented_network_nodes$geometry, pch = 16, add = TRUE)
}
