% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-segmented-network-nodes.R
\name{extract_segmented_network_nodes}
\alias{extract_segmented_network_nodes}
\title{Extract nodes of segmented network from a road network}
\usage{
extract_segmented_network_nodes(road_network, segment_length)
}
\arguments{
\item{road_network}{A \code{road_network} object created with \code{create_network()}.}

\item{segment_length}{The length of each segment to sample along the links.}
}
\value{
An \code{sf} object with the sampled points.
}
\description{
This function extracts nodes at regular intervals along each link in
a road network
}
\examples{
# Create a road network
road_network <- create_road_network(demo_roads)

# Extract nodes with a segment length of 1
segmented_network_nodes <- extract_segmented_network_nodes(road_network, 1)
segmented_network_nodes

# Plot the segmented network nodes
plot(road_network)
plot(segmented_network_nodes$geometry, add = TRUE, pch = 16, col = "#E69F00")
}
