% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-duration.R
\name{find_duration}
\alias{find_duration}
\title{Find the index of the duration that contains the event time.}
\usage{
find_duration(events, durations, ...)
}
\arguments{
\item{events}{A spatiotemporal event object.}

\item{durations}{A duration object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A vector of indices corresponding to the durations that contain
the event times.
}
\description{
This function find the index of the duration that contains the event time.
}
\examples{
# Create a spatiotemporal event object
accidents <- create_spatiotemporal_events(sample_accidents)

# Create a duration object
durations <- create_durations("1 hour")

# Find the duration that contains each accident
find_duration(accidents, durations)
}
