% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate unique IDs based on parent IDs}
\usage{
generate_ids(parent_list, id_format)
}
\arguments{
\item{parent_list}{A list of parent IDs.}

\item{id_format}{A format string for the IDs.}
}
\value{
A vector of unique IDs.
}
\description{
This function generates unique IDs based on parent IDs.
It first extracts the first ID from each parent ID,
then ranks these IDs and uses the ranks to format the IDs.
}
\examples{
parent_list <- c("rd_0001", "rd_0002", "rd_0003")
id_format <- "jn_\%06x"
generate_ids(parent_list, id_format)
}
