% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-start-point.R
\name{get_start_point}
\alias{get_start_point}
\title{Get the start point of a "LINESTRING" or "MULTILINESTRING"}
\usage{
get_start_point(x)
}
\arguments{
\item{x}{A "LINESTRING" or "MULTILINESTRING" object.}
}
\value{
For "LINESTRING" input, returns a "POINT" object representing the start
point.
For "MULTILINESTRING" input, returns a "MULTIPOINT" object representing
the start points of each "LINESTRING".
}
\description{
This function extracts the start point of a "LINESTRING" or
"MULTILINESTRING"
}
\examples{
# Example for "LINESTRING"
linestring <- structure(matrix(
  c(0, 0, 1, 1, 2, 1),
  ncol = 2, byrow = TRUE
), class = c("XY", "LINESTRING", "sfg"))
start_point <- get_start_point(linestring)
plot(linestring)
plot(start_point, add = TRUE)

# Example for "MULTILINESTRING"
multilinestring <- structure(list(
  matrix(c(0, 0, 1, 1), ncol = 2, byrow = TRUE),
  matrix(c(2, 1, 3, 2), ncol = 2, byrow = TRUE)
), class = c("XY", "MULTILINESTRING", "sfg"))
start_points <- get_start_point(multilinestring)
plot(multilinestring)
plot(start_points, add = TRUE)
}
\seealso{
\code{\link[=get_start_point]{get_start_point()}} for obtaining the start point
}
