% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-linestring-endpoint.R
\name{is_linestring_endpoint}
\alias{is_linestring_endpoint}
\title{Check if a point is a start or end point of linestring}
\usage{
is_linestring_endpoint(point, linestring)
}
\arguments{
\item{point}{A point object.}

\item{linestring}{A linestring object.}
}
\value{
\code{TRUE} if the point is start or end point of the linestring,
\code{FALSE} otherwise.
}
\description{
Check if a point is a start or end point of linestring
}
