% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-osm-highway-values.R
\docType{data}
\name{osm_highway_values}
\alias{osm_highway_values}
\title{Values for the 'highway' tag in OpenStreetMap data}
\format{
A character vector of length 12.
}
\usage{
osm_highway_values
}
\description{
\code{osm_highway_values} contains a character vector of values for the 'highway'
tag in OpenStreetMap data.
}
\examples{
osm_highway_values
}
\keyword{datasets}
