% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavement-package.R
\docType{package}
\name{pavement-package}
\alias{pavement}
\alias{pavement-package}
\title{pavement: Analyzing Spatial Events on Roadways}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Pavement is a package designed to analyze spatial events occurring on roadways. It provides a comprehensive toolkit for working with spatial data, empowering users to understand patterns and trends in road-related phenomena.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NONONOexe/pavement}
  \item \url{https://nononoexe.github.io/pavement/}
  \item Report bugs at \url{https://github.com/NONONOexe/pavement/issues}
}

}
\author{
\strong{Maintainer}: Keisuke Ando \email{ando@maslab.aitech.ac.jp}

Other contributors:
\itemize{
  \item Takeshi Uchitane [thesis advisor]
  \item Naoto Mukai [thesis advisor]
  \item Kazunori Iwata [thesis advisor]
  \item Nobuhiro Ito [thesis advisor]
  \item Yong Jiang [thesis advisor]
}

}
\keyword{internal}
