% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-road-network.R
\name{plot.road_network}
\alias{plot.road_network}
\title{Plot a road network}
\usage{
\method{plot}{road_network}(x, y, mode = c("default", "event", "graph"), ...)
}
\arguments{
\item{x}{A road network object.}

\item{y}{This argument is not used.}

\item{mode}{The plotting mode. "event" mode shows the location events
assigned to the road network. "graph" mode shows the direction of links
when the network is directed. This utilizes the \code{plot.igraph} function.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
This function plots a road network.
}
\examples{
# Create the road network
road_network <- create_road_network(sample_roads, events = sample_accidents)

# Plot the road network
plot(road_network)

# Plot the road network with events
plot(road_network, mode = "event")

# Plot the road network as a graph
plot(road_network, mode = "graph")
}
