% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-tnkde-data.R
\name{prepare_tnkde_data}
\alias{prepare_tnkde_data}
\title{Prepare a segmented network for TNKDE}
\usage{
prepare_tnkde_data(segmented_network, events_sf)
}
\arguments{
\item{segmented_network}{A segmented network (output of create_segmented_network()).}

\item{events_sf}{An sf object of events with geometry and time.}
}
\value{
segmented_network with \verb{$events} component and segment-level counts
}
\description{
Assign events to the nearest road segments and count them.
}
