% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-spatiotemporal-event.R
\name{print.spatiotemporal_event}
\alias{print.spatiotemporal_event}
\title{Print a spatiotemporal event collection}
\usage{
\method{print}{spatiotemporal_event}(x, ...)
}
\arguments{
\item{x}{A \code{spatiotemporal_event} object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The \code{spatiotemporal_event} object, invisibly.
}
\description{
This function prints a summary of a spatiotemporal event collection object.
}
\examples{
spatiotemporal_events <- create_spatiotemporal_event(sample_accidents)
print(spatiotemporal_events)
}
