% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-points-near-endpoints.R
\name{remove_points_near_endpoints}
\alias{remove_points_near_endpoints}
\title{Remove points near endpoints of a linestring}
\usage{
remove_points_near_endpoints(points, linestring, tolerance = 0.01)
}
\arguments{
\item{points}{A \code{sfc} object containing points to filter.}

\item{linestring}{A linestring object.}

\item{tolerance}{A numeric value representing the maximum distance
from the line's ends to consider points for filtering. Points within
this distance will be excluded.}
}
\value{
A \code{sfc} object containing the filtered points.
}
\description{
This function removes points from a set of points that are within a
specified tolerance distance of the start or end point of a linestring.
}
\examples{
library(sf)

# Create a points
points <- st_sfc(
  st_point(c(0.000, 1.000)),
  st_point(c(0.500, 0.600)),
  st_point(c(1.000, 0.010)),
  st_point(c(1.500, 0.501)),
  st_point(c(2.000, 0.990))
)

# Create a linestring
linestring <- create_linestring(0, 1, 1, 0, 2, 1)

# Plot the points
plot(linestring, col = "gray")
plot(points, add = TRUE)

# Filter points within a tolerance distance (default: 0.01)
filtered_points <- remove_points_near_endpoints(points, linestring)

# Plot the filtered points
plot(linestring, col = "gray")
plot(filtered_points, add = TRUE)
}
