% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample-accidents-multiyear.R
\docType{data}
\name{sample_accidents_multiyear}
\alias{sample_accidents_multiyear}
\title{Sample accident data for five years}
\format{
An \code{sf} object with 50 rows and 6 variables:
\describe{
\item{id}{Unique accident identifier (character).}
\item{time}{Time of the accident (integer, hour of day).}
\item{weather}{Weather condition (factor: "Sunny", "Cloudy", "Rainy", "Foggy", "Snowy").}
\item{severity}{Severity of accident (factor: "Minor", "Serious", "Fatal").}
\item{geometry}{Spatial point location (\code{sfc_POINT}).}
\item{year}{Year of occurrence (integer, 1–5).}
}
}
\usage{
sample_accidents_multiyear
}
\description{
Synthetic traffic accident data over five years.
}
\examples{
sample_accidents_multiyear
}
\keyword{datasets}
