% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-points-along-linestring.R
\name{sample_points_along_linestring}
\alias{sample_points_along_linestring}
\title{Sample points along a linestring}
\usage{
sample_points_along_linestring(linestrings, segment_length)
}
\arguments{
\item{linestrings}{A \code{sfc} object containing linestrings.}

\item{segment_length}{The desired length of each segment between sampled
points.}
}
\value{
A \code{sfc} object containing the sampled points, grouped as MULTIPOINTs
for each input linestring.
}
\description{
This function samples points at regular intervals along a linestring object.
}
\examples{
library(sf)

# Create a linestrings
linestrings <- c(
  create_linestring(0, 1, 2, 1),
  create_linestring(1, 1.3, 1, 0, 2, 0.5)
)

# Sample points along the linestrings
sampled_points <- sample_points_along_linestring(linestrings, 0.5)

# Plot the sampled points
plot(linestrings)
plot(sampled_points, add = TRUE, pch = 16, col = c("#E69F00", "#56B4E9"))
}
