% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample-roads.R
\docType{data}
\name{sample_roads}
\alias{sample_roads}
\title{Sample roads dataset}
\format{
A \code{sf} linestring object
}
\usage{
sample_roads
}
\description{
\code{sample_roads} a sample dataset representing simple roads.
It intended for demonstration and testing.
}
\examples{
# Show information about the demonstration roads
sample_roads

# Shape of the demonstration roads
plot(sample_roads$geometry)
}
\keyword{datasets}
