% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-events.R
\name{set_events}
\alias{set_events}
\alias{set_events.road_network}
\alias{set_events.segmented_network}
\alias{set_events.spatiotemporal_network}
\title{Set events on a space.}
\usage{
set_events(x, events, ...)

\method{set_events}{road_network}(x, events, ...)

\method{set_events}{segmented_network}(x, events, ...)

\method{set_events}{spatiotemporal_network}(x, events, ...)
}
\arguments{
\item{x}{A \code{road_network} or \code{segmented_network} object.}

\item{events}{A \code{sf} object representing events.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The input object with the events added.
}
\description{
This function sets events on a space.
}
\examples{
# Create the road network
road_network <- create_road_network(sample_roads)

# Set accidents on the road network
road_network <- set_events(road_network, sample_accidents)
}
