% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-spatiotemporal-network.R
\name{spatiotemporal_network}
\alias{spatiotemporal_network}
\alias{create_spatiotemporal_network}
\title{Create spatiotemporal network}
\usage{
create_spatiotemporal_network(
  road_network,
  spatial_length = 1,
  temporal_length = "1 hour",
  events = NULL,
  ...
)
}
\arguments{
\item{road_network}{A \code{road_network} object.}

\item{spatial_length}{The spatial length of segment.}

\item{temporal_length}{The temporal length of segment.}

\item{events}{A event collection object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A \code{spatiotemporal_network} object.
}
\description{
This function creates a spatiotemporal network from a road network.
}
\examples{
# Create a road network
road_network <- create_road_network(sample_roads)

# Create a spatiotemporal road network
spatiotemporal_network <- create_spatiotemporal_network(
  road_network,
  spatial_length = 0.5,
  temporal_length = "4 hour"
)
spatiotemporal_network

# Plot the spatiotemporal road network
plot(spatiotemporal_network)
}
