% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-linestrings.R
\name{split_linestrings}
\alias{split_linestrings}
\title{Split multiple linestrings into segments}
\usage{
split_linestrings(linestrings, split_points, tolerance = 0.01)
}
\arguments{
\item{linestrings}{An \code{sfc} object containing \code{LINESTRING} geometries,
or a single linestring geometry.}

\item{split_points}{An \code{sfc} object containing \code{MULTIPOINT} geometries,
(one per linestring), or a set of \code{POINT}/\code{MULTIPOINT} geometries for a
single \code{LINESTRING}.}

\item{tolerance}{A numeric value representing the maximum distance
allowed between a linestring and its split points. Points beyond this
tolerance will be ignored.}
}
\value{
If the input is a single \code{LINESTRING}, returns an \code{sfc} of
\code{LINESTRING}s. If the input is multiple \code{LINESTRING}s, returns a list of
\code{sfc} objects, one for each input \code{LINESTRING}.
}
\description{
Split multiple linestrings into segments
}
\examples{
# Single `LINESTRING`
line   <- create_linestring(0, 0, 0, 2, 2, 2, 2, 1, 0, 1)
points <- create_points(0, 1, 1, 2, 2, 1, as_multipoint = TRUE)
plot(line)
plot(points, add = TRUE)

segments <- split_linestrings(line, points)[[1]]
plot(segments, col = c("#E69F00", "#56B4E9", "#009E73", "#F0E442"), lwd = 2)

# Multiple `LINESTRING`s
lines  <- c(create_linestring(0.0, 0.0, 0.0, 2.0, 2.0, 2.0, 2.0, 1.0, 0.0,
                              1.0),
            create_linestring(2.5, 1.0, 3.5, 1.0, 3.5, 0.0, 2.5, 0.0, 2.5,
                              0.5, 3.5, 0.5))
points <- c(create_points(0.0, 1.0, 1.0, 2.0, 2.0, 1.0,
                          as_multipoint = TRUE),
            create_points(2.5, 0.5,
                          as_multipoint = TRUE))
plot(lines)
plot(points, add = TRUE)

segments <- do.call(c, split_linestrings(lines, points))
plot(segments,
     col = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
             "#D55E00"),
     lwd = 2)

}
