% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-bbox.R
\name{create_bbox}
\alias{create_bbox}
\title{Create a bounding box}
\usage{
create_bbox(x)
}
\arguments{
\item{x}{A named vector containing the bounding box coordinates.}
}
\value{
A 2x2 matrix representing the bounding box.
}
\description{
This function creates a bounding box from either cardinal coordinates
(north, south, east and west) or center coordinates and dimensions
(center longitude and latitude, and size of width and height).
}
