% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-linestring.R
\name{create_linestring}
\alias{create_linestring}
\title{Create a linestring geometry}
\arguments{
\item{...}{A series of x, y coordinates.}

\item{coordinates}{A \code{coordinates} object.}

\item{crs}{The coordinate reference system of the points.}
}
\value{
A simple feature linestring object.
}
\description{
This function creates a simple feature linestring object
from a series of x, y coordinates.
}
\examples{
# Create a linestring from individual coordinates
linestring_1 <- create_linestring(0, 1, 1, 0, 2, 1)
linestring_1
plot(linestring_1)

# Create a linestring from a numeric vector
linestring_2 <- create_linestring(
  c(0, 1, 1, 1, 1, 0, 0, 0, 0, 0.5, 1, 0.5)
)
linestring_2
plot(linestring_2)
}
