% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-points.R
\name{create_points}
\alias{create_points}
\alias{create_points.numeric}
\alias{create_points.coordinates}
\title{Create points geometry}
\usage{
create_points(...)

\method{create_points}{numeric}(..., crs = NULL)

\method{create_points}{coordinates}(coordinates, crs = NULL, ...)
}
\arguments{
\item{...}{A series of x, y coordinates.}

\item{crs}{The coordinate reference system.}

\item{coordinates}{A \code{coordinates} object.}
}
\value{
A simple feature points object.
}
\description{
This function creates a simple feature collection of points
from a series of x, y coordinates.
}
\examples{
# Create a points from individual coordinates
points_1 <- create_points(0, 1, 1, 0, 2, 1)
points_1
plot(points_1)

# Create a linestring from a numeric vector
points_2 <- create_points(
  c(0, 1, 1, 1, 1, 0, 0, 0, 0, 0.5, 1, 0.5)
)
points_2
plot(points_2)
}
