% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-points.R
\name{create_points}
\alias{create_points}
\title{Create a simple feature collection of points}
\arguments{
\item{...}{A series of x, y coordinates.}

\item{coordinates}{A \code{coordinates} object.}

\item{crs}{The coordinate reference system of the points.}
}
\value{
A simple feature points object.
}
\description{
This function creates a simple feature collection of points
from a series of x, y coordinates.
}
\examples{
# Create points
points <- create_points(0, 1, 1, 0, 2, 1)

# Plot the points
plot(points)
}
