% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-polygon.R
\name{create_polygon}
\alias{create_polygon}
\alias{create_polygon.numeric}
\alias{create_polygon.coordinates}
\title{Create a polygon geometry}
\usage{
create_polygon(...)

\method{create_polygon}{numeric}(..., crs = NULL)

\method{create_polygon}{coordinates}(coordinates, crs = NULL, ...)
}
\arguments{
\item{...}{A series of x, y coordinates.}

\item{crs}{The coordinate reference system.}

\item{coordinates}{A \code{coordinates} object.}
}
\value{
A polygon object.
}
\description{
This function creates a polygon from a bounding box or a series of
x, y coordinates.
}
\examples{
polygon <- create_polygon(
  136.9009, 35.16377,
  136.9159, 35.16377,
  136.9159, 35.17377,
  136.9009, 35.17377,
  crs = 4326
)
}
