% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude-points.R
\name{exclude_points}
\alias{exclude_points}
\title{Exclude points from a points object}
\usage{
exclude_points(points, points_to_exclude)
}
\arguments{
\item{points}{A \code{sfc} object containing points.}

\item{points_to_exclude}{A \code{sfc} object containing points to exclude.}
}
\value{
A \code{sfc} object containing points with the excluded points removed.
}
\description{
This function exclude points from a points object.
}
