% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-connected-links.R
\name{get_connected_links}
\alias{get_connected_links}
\title{Get connected links from a node}
\usage{
get_connected_links(network, node_id)
}
\arguments{
\item{network}{A road network object.}

\item{node_id}{The ID of the node.}
}
\value{
A data frame containing the connected links.
}
\description{
This function retrieves all links connected to a specified node in a road
network.
}
\examples{
# Create a road network
road_network <- create_road_network(demo_roads)
target_node <- road_network$nodes[3,]
connected_links <- get_connected_links(road_network, target_node$id)

# Plot the target node and the connected links
plot(road_network, col = "gray")
plot(connected_links$geometry, add = TRUE, col = "#D55E00", lwd = 2)
plot(target_node$geometry, add = TRUE, pch = 19)
}
