% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-local-moran.R
\name{plot_local_moran}
\alias{plot_local_moran}
\title{Plot Spatio-temporal Local Moran's I results}
\usage{
plot_local_moran(moran_results_object)
}
\arguments{
\item{moran_results_object}{The output object from \code{calculate_local_moran},
which contains a \verb{$moran_results} data frame. The function can handle
objects of class \code{segmented_network} or \code{spatiotemporal_network}.}
}
\value{
A \code{plotly} object representing the interactive 3D plot.
}
\description{
Creates an interactive 3D plot to visualize the results of the
\code{calculate_local_moran} function, showing significant spatio-temporal
clusters (e.g., High-High, Low-Low) in a space-time cube.
}
