% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample-roads.R
\docType{data}
\name{sample_roads}
\alias{sample_roads}
\title{Sample roads data}
\format{
A \code{sf} linestring object with 6 rows and 2 columns:
\describe{
\item{id}{Road ID}
\item{layer}{Layer of the road (indicates a road on an elevated
structure such as a bridge, tunnel, or highway)}
\item{geometry}{Linestring geometry of the road}
}
}
\usage{
sample_roads
}
\description{
A sample dataset representing simple roads.
It intended for demonstration and testing and does not represent
real-world roads.
}
\examples{
# Show information about the roads
sample_roads

# Shape of the roads
plot(sample_roads$geometry)
}
\keyword{datasets}
