% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-events.R
\name{set_events}
\alias{set_events}
\title{Set events on a space.}
\usage{
set_events(x, events, ...)
}
\arguments{
\item{x}{A space object.}

\item{events}{A sf or spatiotemporal events object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The input object with the events added.
}
\description{
This function sets events on a space.
}
\examples{
# Create the road network
road_network <- create_road_network(sample_roads)

# Set accidents on the road network
road_network <- set_events(road_network, sample_accidents)
}
