% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-spatio-properties.R
\name{test_spatio_properties}
\alias{test_spatio_properties}
\title{Test for temporal consistency of spatio-temporal properties}
\usage{
test_spatio_properties(list_of_moran_results, p_value_threshold = 0.05)
}
\arguments{
\item{list_of_moran_results}{A list where each element is the output of
\code{calculate_local_moran} for a specific year.}

\item{p_value_threshold}{The p-value cutoff to determine consistency (default: 0.05).}
}
\value{
A data frame summarizing the consistency test for each spatiotemporal_segement.
}
\description{
This function evaluates the temporal consistency of Local Moran's I cluster
classifications for each spatio-temporal unit (spatiotemporal_segement) across multiple years.
}
