% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-to-cartesian.R
\name{transform_to_cartesian}
\alias{transform_to_cartesian}
\title{Transform to Cartesian coordinates}
\usage{
transform_to_cartesian(x)
}
\arguments{
\item{x}{A spatial object.}
}
\value{
A spatial object transformed to Cartesian coordinates.
}
\description{
This function transforms a spatial object to Cartesian coordinates (plane
rectangular coordinate system). Currently, it supports the coordinate
system used in Aichi Prefecture (JDG2011, EPSG:6675).
}
\seealso{
\code{\link[=transform_to_geographic]{transform_to_geographic()}} to transform to geographic coordinates.
}
