% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-to-geographic.R
\name{transform_to_geographic}
\alias{transform_to_geographic}
\title{Transform to geographic coordinates}
\usage{
transform_to_geographic(x)
}
\arguments{
\item{x}{A spatial object.}
}
\value{
A spatial object transformed to geographic coordinates.
}
\description{
This function transforms a spatial object to geographic coordinates
(WGS84, EPSG:4326).
}
\examples{
# Create points
points <- create_points(-24073.54, -92614.18, crs = 6675)
points

# Transform to geographic coordinates
transformed <- transform_to_geographic(points)
transformed
}
\seealso{
\code{\link[=transform_to_cartesian]{transform_to_cartesian()}} to transform to Cartesian coordinates.
}
